# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  template_internal_http_datatransfers_requests.EIDPollRequest:
    properties:
      google_link_token:
        description: |-
          GoogleLinkToken нь Google-ээр эхний удаа нэвтэрсэн хэрэглэгч eID-ээр
          баталгаажуулж байгаа үед л ирнэ (сонголттой).
        type: string
      session_id:
        type: string
    required:
    - session_id
    type: object
  template_internal_http_datatransfers_requests.EIDStartByNationalIDRequest:
    properties:
      callbackUrl:
        description: |-
          CallbackUrl (сонголт): SAME-DEVICE (утасны browser) үед <origin>/auth/eid/callback; хоосон
          бол CROSS-DEVICE (desktop). Backend force-normalize хийнэ.
        type: string
      national_id:
        type: string
    required:
    - national_id
    type: object
...
  template_internal_http_datatransfers_requests.RegisterRequest:
    properties:
      email:
        maxLength: 50
        type: string
      first_name:
        description: нэр (монгол)
        maxLength: 50
        minLength: 1
        type: string
      first_name_en:
        description: нэр (англи)
        maxLength: 50
        type: string
      last_name:
        description: овог (монгол)
        maxLength: 50
        minLength: 1
        type: string
      last_name_en:
        description: овог (англи)
        maxLength: 50
        type: string
      password:
        maxLength: 72
        minLength: 12
        type: string
      username:
        maxLength: 25
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
...
  template_internal_http_datatransfers_requests.SendOTPRequest:
    properties:
      email:
        maxLength: 50
        type: string
    required:
    - email
    type: object
...
  template_internal_http_datatransfers_responses.EIDCertificate:
    properties:
      issuer:
        type: string
      key_type:
        type: string
      not_after:
        type: string
      not_before:
        type: string
      serial:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.EIDInfo:
    properties:
      certificate:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDCertificate'
      civil_id:
        type: string
      document_number:
        type: string
      kyc_level:
        description: сертификатын түвшин
        type: string
      national_id:
        description: регистрийн дугаар
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.EIDPollResponse:
    properties:
      created_at:
        type: string
      eid:
        allOf:
        - $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDInfo'
        description: |-
          EID нь eID-ээр нэвтэрсэн хэрэглэгчийн identity + сертификатын мэдээлэл.
          Нууц үгээр бүртгүүлсэн хэрэглэгчид nil (omitempty).
      eid_proxy:
        description: |-
          EIDProxy нь SSO eID proxy идэвхтэй эсэхийг заана — идэвхтэй бол иргэн
          локал eID linkage-гүй (SSO-ээр нэвтэрсэн) байсан ч eID PKI самбарыг SSO-
          гоор дамжуулан үзэж болно. Frontend eID хуудсуудыг үүгээр нээнэ.
        type: boolean
      email:
        type: string
      first_name:
        type: string
      first_name_en:
        type: string
      full_name:
        type: string
      full_name_en:
        type: string
      google:
        allOf:
        - $ref: '#/definitions/template_internal_http_datatransfers_responses.GoogleInfo'
        description: |-
          Google нь холбогдсон Google account-аас хадгалсан профайл. Google
          холбоогүй хэрэглэгчид nil (omitempty).
      id:
        type: string
      last_name:
        type: string
      last_name_en:
        type: string
      mfa_required:
        type: boolean
      mfa_token:
        type: string
      refresh_token:
        type: string
      role_id:
        type: integer
      state:
        type: string
      token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  template_internal_http_datatransfers_responses.EIDStartResponse:
    properties:
      device_link_url:
        type: string
      expires_at:
        type: string
      session_id:
        type: string
      verification_code:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.GoogleInfo:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      linked_at:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.UserResponse:
    properties:
      created_at:
        type: string
      eid:
        allOf:
        - $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDInfo'
        description: |-
          EID нь eID-ээр нэвтэрсэн хэрэглэгчийн identity + сертификатын мэдээлэл.
          Нууц үгээр бүртгүүлсэн хэрэглэгчид nil (omitempty).
      eid_proxy:
        description: |-
          EIDProxy нь SSO eID proxy идэвхтэй эсэхийг заана — идэвхтэй бол иргэн
          локал eID linkage-гүй (SSO-ээр нэвтэрсэн) байсан ч eID PKI самбарыг SSO-
          гоор дамжуулан үзэж болно. Frontend eID хуудсуудыг үүгээр нээнэ.
        type: boolean
      email:
        type: string
      first_name:
        type: string
      first_name_en:
        type: string
      full_name:
        type: string
      full_name_en:
        type: string
      google:
        allOf:
        - $ref: '#/definitions/template_internal_http_datatransfers_responses.GoogleInfo'
        description: |-
          Google нь холбогдсон Google account-аас хадгалсан профайл. Google
          холбоогүй хэрэглэгчид nil (omitempty).
      id:
        type: string
      last_name:
        type: string
      last_name_en:
        type: string
      refresh_token:
        type: string
      role_id:
        type: integer
      token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  template_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Government Developer Portal V3.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Government Developer Portal V3.0 API
  version: "1.0"
paths:
...
  /auth/eid/poll:
    post:
      consumes:
      - application/json
      description: session_id-ийн төлвийг IdP-ээс long-poll-оор (≤25с) асууна. state
        нь RUNNING/COMPLETE/EXPIRED/REFUSED. COMPLETE үед identity-аар хэрэглэгчийг
        бүртгэж/шинэчилж, access+refresh токен хосыг /login-той ижил хэлбэрээр буцаана.
      parameters:
      - description: eID session id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.EIDPollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Session state (with tokens if COMPLETE)
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDPollResponse'
              type: object
        "400":
          description: Malformed JSON body or missing session_id
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "500":
          description: Failed to reach eID provider
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: eID session-ийн төлвийг асуух (long-poll)
      tags:
      - auth
  /auth/eid/start:
    post:
      consumes:
      - application/json
      description: Гадаад eID identity provider дээр QR/deep-link нэвтрэлтийг эхлүүлж,
        клиент харуулах session мэдээллийг (session_id, device_link_url, verification_code,
        expires_at) буцаана. Дараа нь /auth/eid/poll руу session_id-г дамжуулж төлвийг
        асууна.
      produces:
      - application/json
      responses:
        "200":
          description: eID session started
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDStartResponse'
              type: object
        "500":
          description: Failed to reach eID provider
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: eID нэвтрэлт эхлүүлэх (QR / deep-link)
      tags:
      - auth
  /auth/eid/start-id:
    post:
      consumes:
      - application/json
      description: Иргэний РД (national_id)-аар нэвтрэлтийг эхлүүлж, тухайн РД-тэй
        холбоотой бүртгэлтэй төхөөрөмж рүү баталгаажуулах prompt push хийлгэнэ. QR/device_link
        шаардлагагүй тул зөвхөн session_id, verification_code, expires_at буцна. Дараа
        нь /auth/eid/poll руу session_id-г дамжуулж төлвийг асууна.
      parameters:
      - description: Иргэний РД
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.EIDStartByNationalIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: eID session started
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDStartResponse'
              type: object
        "400":
          description: Malformed JSON body or missing national_id
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "500":
          description: Failed to reach eID provider
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: eID нэвтрэлт эхлүүлэх (РД-аар push)
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Хэрэглэгчийн бүртгэлийг идэвхгүй төлөвт үүсгэнэ. Дуудагч нь /auth/login
        амжилттай болохоос өмнө бүртгэлийг идэвхжүүлэхийн тулд /auth/send-otp + /auth/verify-otp-г
        дагах ёстой.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.UserResponse'
              type: object
        "400":
          description: Malformed JSON body
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "409":
          description: Email or username already in use
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error (per-field detail in data.errors)
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: Шинэ хэрэглэгч бүртгэх
      tags:
      - auth
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: GeregeCloud Verify API-аар OTP илгээж, буцаасан request_id-г TTL-тэйгээр
        Redis-д хадгална. Кодын үүсгэлт/хадгалалт/илгээлтийг Verify үйлчилгээ хариуцна.
      parameters:
      - description: Email to send OTP to
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP enqueued
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Account already activated
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Email not registered
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "500":
          description: Failed to enqueue mail
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: Хэрэглэгчийн email рүү OTP код илгээх
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"